Version 1.1
New variable g:scaleFontExec_{MODE}, the content of which will be executed before reconfiguring the window geometry. This can be used to set guioptions as needed (e.g., for a full screen mode without menu bar and all as for NormalFull). If this variable isn't set, the variable g:scaleFontExec_0 will be executed that resets the guioptions to its initial value.  ScaleFontMaximizeWindow() is defined for OS other than Windows too (kind of). Build modes menu (see g:scaleFontMenuPrefix).